apt 명칭 Apartment 수정, apiaptController 에서 전부 받던 거 Service 폴더로 이동
