+++ src/main/java/com/ap4j/bma/service/apartment/ApartmentService.java
@@ -1,0 +1,10 @@
+package com.ap4j.bma.service.apt;
+
+import com.ap4j.bma.model.entity.apt.AptDTO;
+
+import java.util.ArrayList;
+
+public interface AptService {
+
+	public ArrayList<AptDTO> getAptLists();
+}
